+++ NEWS.md
@@ -2,8 +2,8 @@
 
 * Setup a `cplot.default()` method and modified documentation of `cplot()`, `image()`, and `persp()` methods slightly. (#84, h/t Luke Sonnet)
 * Improve the documentation the behavior of `cplot()` for generalized linear models, which can generate unexpected confidence intervals (albeit ones consistent with base R's behavior). (#92)
-* Models fit using the `lme4` package can now have variance estimation via 
-bootstrap and simulation (#105).
+* Fix bug that caused spurious `NA`s and errors in `margins()` when `vce` was `"bootstrap"` or `"simulation"` and `variables` had a length of 1. (#112)
+* Models fit using the `lme4` package can now have variance estimation via bootstrap and simulation (#105).
 
 # margins 0.3.24
 
@@ -44,7 +44,6 @@ bootstrap and simulation (#105).
 * Updated examples in `README.Rmd`. (#83)
 
 # margins 0.3.16
->>>>>>> master
 
 * Fixed a bug in `cplot()` when `xvar` was of class "ordered". (#77, h/t Francisco Llaneras)
 * Fixed a bug in `plot.margins()` when `at` contained only one variable. (#78, h/t @cyberbryce)